サーバーの情報からホーム画面の表示
