+++ src/components/AllMap.jsx
@@ -27,7 +27,7 @@ const AllMap = () => {
   const center = [34.699228, 135.494274];
 
   return (
-    <MapContainer center={center} zoom={16} scrollWheelZoom={true} style={{ height: '500px', width: '100%' }}>
+    <MapContainer center={center} zoom={16} scrollWheelZoom={true} style={{ height: '700px', width: '100%' }}>
       <TileLayer
         attribution='&copy; <a href="https://www.openstreetmap.org/">OpenStreetMap</a> contributors'
         url='https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
